requirements.txt
discord.py
praw
aiml
requests
psutil
rule34
python-dotenv
pymongo
dnspython